# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bitbiz_hias-core_domains_claims_schema.BulkSubmitClaimsRequest:
    properties:
      claims:
        items:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.SubmitClaimRequest'
        minItems: 1
        type: array
    required:
    - claims
    type: object
...
  github_com_bitbiz_hias-core_domains_claims_schema.LineItemRequest:
    properties:
      diagnosis_code:
        type: string
      procedure_code:
        type: string
      procedure_name:
        type: string
      quantity:
        minimum: 1
        type: integer
      unit_price:
        minimum: 1
        type: integer
    required:
    - procedure_code
    - procedure_name
    - quantity
    - unit_price
    type: object
  github_com_bitbiz_hias-core_domains_claims_schema.ReviewClaimRequest:
    properties:
      decision:
        enum:
        - APPROVED
        - REJECTED
        type: string
      reason:
        type: string
    required:
    - decision
    type: object
  github_com_bitbiz_hias-core_domains_claims_schema.SubmitClaimRequest:
    properties:
      admission_date:
        type: string
      claim_type:
        type: string
      diagnosis_codes:
        items:
          type: string
        type: array
      discharge_date:
        type: string
      line_items:
        items:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.LineItemRequest'
        minItems: 1
        type: array
      member_id:
        type: string
      notes:
        type: string
      policy_id:
        type: string
      preauth_id:
        type: string
      provider_id:
        type: string
      service_date:
        type: string
    required:
    - diagnosis_codes
    - line_items
    - member_id
    - policy_id
    - provider_id
    - service_date
    type: object
...
  github_com_bitbiz_hias-core_domains_claims_schema.VetClaimRequest:
    properties:
      notes:
        type: string
      vetted_amount:
        minimum: 0
        type: integer
    required:
    - vetted_amount
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Health Insurance Administration System - Core API
  title: HIAS Core API
  version: "1.0"
paths:
...
  /api/v1/claim-documents/{id}:
    delete:
      description: Delete a document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a claim document
      tags:
      - Claims
  /api/v1/claims:
    get:
      description: List claims with optional filtering by status, provider, date range,
        or search query
      parameters:
      - description: Filter by claim status
        in: query
        name: status
        type: string
      - description: Filter by provider ID
        in: query
        name: provider
        type: string
      - description: Filter from date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: Filter to date (RFC3339)
        in: query
        name: date_to
        type: string
      - description: Search claim number or status
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List claims
      tags:
      - Claims
    post:
      consumes:
      - application/json
      description: Submit a new insurance claim for processing
      parameters:
      - description: Claim submission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.SubmitClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a new claim
      tags:
      - Claims
  /api/v1/claims/{id}:
    get:
      description: Retrieve a single claim by its unique identifier
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a claim by ID
      tags:
      - Claims
  /api/v1/claims/{id}/approve:
    put:
      description: Approve an insurance claim by its ID
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a claim
      tags:
      - Claims
...
  /api/v1/claims/{id}/documents:
    get:
      description: List all documents associated with a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List claim documents
      tags:
      - Claims
    post:
      consumes:
      - application/json
      description: Upload a document associated with a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Document details (file_name, file_type, file_size, s3_key)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a claim document
      tags:
      - Claims
  /api/v1/claims/{id}/mark-paid:
    put:
      description: Mark a claim as fully paid
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark claim as paid
      tags:
      - Claims
  /api/v1/claims/{id}/mark-part-paid:
    put:
      description: Mark a claim as partially paid
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark claim as partially paid
      tags:
      - Claims
...
  /api/v1/claims/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject an insurance claim by its ID with a reason
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.ReviewClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject a claim
      tags:
      - Claims
  /api/v1/claims/{id}/timeline:
    get:
      description: Returns the status change history (timeline) for a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get claim timeline
      tags:
      - Claims
  /api/v1/claims/{id}/vet:
    put:
      consumes:
      - application/json
      description: Vet an insurance claim by its ID with vetting details
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Vetting details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.VetClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Vet a claim
      tags:
      - Claims
  /api/v1/claims/bulk:
    post:
      consumes:
      - application/json
      description: Submit multiple insurance claims in a single request
      parameters:
      - description: Bulk claim submissions
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_claims_schema.BulkSubmitClaimsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk submit claims
      tags:
      - Claims
  /api/v1/claims/import-csv:
    post:
      consumes:
      - multipart/form-data
      description: Import multiple claims by uploading a CSV file
      parameters:
      - description: CSV file containing claims data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import claims from CSV
      tags:
      - Claims
  /api/v1/claims/sla-breached:
    get:
      description: List all claims that have breached their SLA deadlines
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List SLA-breached claims
      tags:
      - Claims
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"